Fix position after scrolling when a modal's opened
The Datedropper3 remains at the same position relative to the page and does not respect the position of the element it was initialized on, because modal is fixed positioning and Datedropper3 is absolute. We could provide an option for append Datedropper3 to each '.input-prepend' and then I also noticed where it was calculating the position of the Datedropper3 it was taking input.offset()..... , 
so I change that and now it works. ....
Best Regards.